+++ inst/include/MatrixInfo.h
@@ -20,19 +20,34 @@ using namespace std;
 namespace gds {
 
 class MatrixInfo :
-    public DataInfo {
-    public:
-    MatrixInfo() {}
-    MatrixInfo(const vector<string> &ID) :
-        DataInfo(ID) {}
+  public DataInfo {
+  public:
+  MatrixInfo() {}
+  MatrixInfo(const vector<string> &ID) :
+      DataInfo(ID) {}
 
-    void setRowNames( const vector<string> &rn, const int &start, const int &end){
+  void setRowNames( const vector<string> &rn, const int &start, const int &end){
 
-        ID.clear();
-        for(int i=start; i<end; i++){
-            ID.push_back(rn[i]);
-        }
+    ID.clear();
+    for(int i=start; i<end; i++){
+      ID.push_back(rn[i]);
     }
+  }
+
+  /** Extract subset corresponding to idx
+   */
+  MatrixInfo subset(const uvec &idx) const {
+
+    vector<string> ID2;
+    ID2.reserve(idx.n_elem);
+
+    for(int i=0; i<idx.n_elem; i++){
+      ID2.push_back(ID[idx[i]]);
+    }
+
+    return MatrixInfo(ID2);      
+  }
+
 };